Refactor: CurrentUser 애노테이션 처리

+++ backend/bartter/src/main/java/com/ssafy/bartter/community/service/CommunityPostCommentService.java
@@ -30,8 +30,8 @@ public class CommunityPostCommentService {
     /**
      * 동네모임 댓글 작성
      * */
-    public CommunityPostComment createComment(Integer communityPostId, Create request) {
-        User user = userRepository.findById(request.getUserId()).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
+    public CommunityPostComment createComment(Integer communityPostId, Create request, Integer userId) {
+        User user = userRepository.findById(userId).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
         CommunityPost post = communityPostRepository.findById(communityPostId).orElseThrow(() -> new CustomException(ErrorCode.COMMUNITY_POST_NOT_FOUND));
 
         CommunityPostComment comment = new CommunityPostComment(request.getContent());
@@ -44,8 +44,11 @@ public class CommunityPostCommentService {
     /**
      * 동네모임 댓글 삭제
      * */
-    public void deleteComment(Integer commentId) {
+    public void deleteComment(Integer userId, Integer commentId) {
         CommunityPostComment comment = communityPostCommentRepository.findById(commentId).orElseThrow(() -> new CustomException(ErrorCode.COMMUNITY_POST_COMMENT_NOT_FOUND));;
+        if (!comment.getUser().getId().equals(userId)) {
+            throw new CustomException(ErrorCode.UNAUTHORIZED);
+        }
         communityPostCommentRepository.delete(comment);
     }
 }
